requirements.txt
aiogram==3.13.1
Babel==2.13.1
librosa==0.10.2.post1
loguru==0.7.2
mistralai==1.1.0
Pillow==9.2.0
Pillow==11.0.0
pymongo==4.7.1
PySoundFile==0.9.0.post1
python-dotenv==1.0.1
pytz==2023.3.post1
soundfile==0.12.1
SpeechRecognition==3.10.1
torchaudio==2.4.1
